% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getunitsandparams.R
\name{getrelatedterms}
\alias{getrelatedterms}
\title{get all related terms of a concept}
\usage{
getrelatedterms(vocid = NA, relation = NA, vocab = NA)
}
\arguments{
\item{vocid}{mandatory parameter, the XML link(s) of which you want the related concepts}

\item{relation}{optional parameter, the type of the retation you want. It takes as values: "narrower", "broader", "related", "sameas"}

\item{vocab}{optional parameter, the vocab of the returned id's. For example: "P01", "P06"}
}
\description{
This function retuns to you all related terms of one or more concept. The output is a vector with the related ids
}
\examples{
getrelatedterms(vocid = "http://vocab.nerc.ac.uk/collection/S25/current/BE007117/", relation = "narrower", vocab = "P01" )
}
