% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{getskossxmldatainfo}
\alias{getskossxmldatainfo}
\title{get skos xml dataterms}
\usage{
getskossxmldatainfo(vocids = NA, vocabs = NA)
}
\arguments{
\item{vocids}{mandatory parameter, the XML link(s)  you want to read the read}

\item{vocabs}{optional parameter, include which vocabs you want to include. Takes multiple vocabs if separated by a "|" (eg. "P01|P06|Q01")}
}
\description{
This function allows to get the information form different concepts stored in skoss xml. It takes a single concept, multiple concepts or a shema. The output a dataframe with the lables and definitions.
}
\examples{
getskossxmldatainfo("http://vocab.nerc.ac.uk/collection/P06/current/UMCU/")
getskossxmldatainfo(c("http://vocab.nerc.ac.uk/collection/P06/current/UMCU/","http://vocab.nerc.ac.uk/collection/P06/current/ULAA/")
getskossxmldatainfo("http://vocab.nerc.ac.uk/collection/S10/current/", vocabs ="S10|S11")
getskossxmldatainfo("http://dd.eionet.europa.eu/vocabulary/biodiversity/eunishabitats/rdf")
}
