% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{leftfrom}
\alias{leftfrom}
\title{left}
\usage{
leftfrom(x, y, n = 0)
}
\arguments{
\item{x}{mandatory parameter, a vector}

\item{y}{mandatory parameter, the text from which you want the left of}
}
\description{
helpfunction simular to left in excel
}
