% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{readskossxml}
\alias{readskossxml}
\title{read skoss xml file}
\usage{
readskossxml(X)
}
\arguments{
\item{vocid}{mandatory parameter, the XML link  you want to read}
}
\description{
This function reads a skoss xml file
}
