% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{richtfrom}
\alias{richtfrom}
\title{right}
\usage{
richtfrom(x, y, n = 0)
}
\arguments{
\item{x}{mandatory parameter, a vector}

\item{y}{mandatory parameter, the text from which you want the rightmost characters of}
}
\description{
helpfunction simular to right in excel
}
