% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{skossxmlinfo}
\alias{skossxmlinfo}
\title{parse skoss xml file}
\usage{
skossxmlinfo(skossxml)
}
\arguments{
\item{skossxml}{mandatory parameter, the output from readskossxml}
}
\description{
This function pases a skoss xml file and provides a dataframe with lables and definitions.
}
