% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getskosxmldata.R
\name{skossxmlrelations}
\alias{skossxmlrelations}
\title{parse skoss xml file2}
\usage{
skossxmlrelations(skossxml)
}
\arguments{
\item{skossxml}{mandatory parameter, the output from readskossxml}
}
\description{
This function pases a skoss xml file and provides the relations
}
